告警清零

+++ rosen/modules/render_service/core/pipeline/rs_render_service_visitor.cpp
@@ -85,8 +85,8 @@ void RSRenderServiceVisitor::PrepareDisplayRenderNode(RSDisplayRenderNode& node)
     }
 
     ScreenRotation rotation = node.GetRotation();
-    uint32_t logicalScreenWidth = node.GetRenderProperties().GetFrameWidth();
-    uint32_t logicalScreenHeight = node.GetRenderProperties().GetFrameHeight();
+    int32_t logicalScreenWidth = static_cast<int32_t>(node.GetRenderProperties().GetFrameWidth());
+    int32_t logicalScreenHeight = static_cast<int32_t>(node.GetRenderProperties().GetFrameHeight());
 
     if (logicalScreenWidth <= 0 || logicalScreenHeight <= 0) {
         logicalScreenWidth = currScreenInfo.width;
